+++ packages/ui-components/.storybook/main.ts
@@ -1,14 +1,14 @@
 import type { StorybookConfig } from '@storybook/react-webpack5';
-import { dirname, join } from 'path';
+import { dirname, join } from 'node:path';
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
 
   addons: [
+    getAbsolutePath('@storybook/addon-actions'),
     getAbsolutePath('@storybook/addon-links'),
     getAbsolutePath('@storybook/addon-essentials'),
     getAbsolutePath('@storybook/addon-interactions'),
-    getAbsolutePath('@storybook/addon-mdx-gfm'),
     getAbsolutePath('@storybook/addon-webpack5-compiler-babel'),
   ],