+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @verdaccio/auth
 
+## 8.0.0-next-8.19
+
+### Patch Changes
+
+- Updated dependencies [6e404a0]
+- Updated dependencies [4e14aff]
+- Updated dependencies [9350431]
+- Updated dependencies [95e6c8f]
+  - verdaccio-htpasswd@13.0.0-next-8.19
+  - @verdaccio/signature@8.0.0-next-8.11
+  - @verdaccio/core@8.0.0-next-8.19
+  - @verdaccio/loaders@8.0.0-next-8.9
+  - @verdaccio/config@8.0.0-next-8.19
+
 ## 8.0.0-next-8.18
 
 ### Patch Changes