+++ packages/core/file-locking/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 13.0.0-next-8.4
+
+### Patch Changes
+
+- 9350431: chore(deps): clean up core modules
+
 ## 13.0.0-next-8.3
 
 ### Patch Changes