+++ packages/loaders/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @verdaccio/loaders
 
+## 8.0.0-next-8.9
+
+### Patch Changes
+
+- 95e6c8f: chore(deps): hooks, loaders, logger
+- Updated dependencies [9350431]
+  - @verdaccio/core@8.0.0-next-8.19
+
 ## 8.0.0-next-8.8
 
 ### Patch Changes