+++ packages/logger/logger-prettify/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @verdaccio/logger-prettify
 
+## 8.0.0-next-8.3
+
+### Patch Changes
+
+- 95e6c8f: chore(deps): hooks, loaders, logger
+
 ## 8.0.0-next-8.2
 
 ### Patch Changes